requirements.txt
Flask
mysql-connector-python
gunicorn
psycopg2-binary